Update App.R (#292)
* Update App.R

I  find the an error when running in the hipot site. It is because when    testing the package, I use the "load_all()" command and it would also load the functions that are not exported.
Now, I have add the predix "UCSCXenaShiny:::" for that function. Next time, I will test package by building it in local environment.

* fix a bug